gatillos Caba ok and rename .json files and dependencias  links modified
