+++ src/data/index.js
@@ -3,11 +3,10 @@ import caba from "../data/caba.json";
 import barriosCaba from "../data/barriosCaba.json";
 import laPlata  from "../data/laPlata.json";
 import departamentosLaPlata  from "../data/laPlata.json";
-import dependenciasLaPlata  from "../data/dependenciasLaPlata.json";
 import dependenciasCaba  from "../data/dependenciasCaba.json";
-import gatillo from "../data/gatilloLaPlata.json";
+import gatillo from "../data/gatilloCaba.json";
 
 
 
 
-export {  departamentos, caba, barriosCaba, laPlata, departamentosLaPlata, dependenciasLaPlata, gatillo, dependenciasCaba };
+export {  departamentos, caba, barriosCaba, laPlata, departamentosLaPlata, gatillo, dependenciasCaba };
