+++ src/main.jsx
@@ -33,9 +33,8 @@ const loader = async () => ({
     caba: "data/caba.json",
     laPlata: "data/laPlata.json",
     departamentosLaPlata: "data/departamentosLaPlata.json",
-    dependenciasLaPlata: "data/dependenciasLaPlata.json",
     dependenciasCaba: "data/dependenciasCaba.json",
-    gatillo: "data/gatilloLaPlata.json",
+    gatillo: "data/gatilloCaba.json",
   }),
 });
